requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
boto3==1.40.25
botocore==1.40.25
certifi==2025.8.3
cfn-flip==1.3.0
charset-normalizer==3.4.3
click==8.2.1
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
hjson==3.1.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
packaging==25.0
placebo==0.9.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
troposphere==4.9.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
zappa==0.60.2
